Use submodules instead of exported namespaces in `@wuespace/telestion`
Previously, anything that wasn't exported on the root level of the `@wuespace/telestion` library had to be imported indirectly through namespaces:

```ts
import { utils } from '@wuespace/telestion';
const { wait } = utils;
```

This didn't provide a great developer experience, as imports couldn't be reoslved automatically.

Instead, these features now must be imported using submodules / path imports:

```ts
import { wait } from '@wuespace/telestion/utils';
```

As of the time of this change, a total of six of these modules exist:
- `@wuespace/telestion/application`, corresponding to `import { application } from '@wuespace/telestion'`
- `@wuespace/telestion/auth`, corresponding to `import { auth } from '@wuespace/telestion'`
- `@wuespace/telestion/nats`, corresponding to `import { nats } from '@wuespace/telestion'`
- `@wuespace/telestion/user-data`, corresponding to `import { userData } from '@wuespace/telestion'`
- `@wuespace/telestion/utils`, corresponding to `import { utils } from '@wuespace/telestion'`
- `@wuespace/telestion/widget`, corresponding to `import { widget } from '@wuespace/telestion'`

The previous namespaces were removed.
