+++ frontend-react/src/lib/application/routes/dashboard.ts
@@ -2,7 +2,7 @@ import { TelestionOptions } from '../model.ts';
 import { isLoggedIn } from '../../auth';
 import { generatePath, redirect } from 'react-router-dom';
 import { getEmptyDashboard, getUserData, setUserData } from '../../user-data';
-import { isUserDataUpToDate } from '../../utils.ts';
+import { isUserDataUpToDate } from '../../utils';
 
 export function dashboardCreateAction({ version }: TelestionOptions) {
 	return () => {
