+++ frontend-react/src/lib/index.ts
@@ -22,13 +22,6 @@
  * @see {@link initTelestion}
  */
 
-export * as application from './application';
-export * as auth from './auth';
-export * as userData from './user-data';
-export * as widget from './widget';
-export * as utils from './utils.ts';
-export * as nats from './nats';
-
 // application
 export { initTelestion, useWidgetConfig } from './application';
 export type { TelestionOptions } from './application';
