+++ frontend-react/src/lib/mod-utils.ts
@@ -1,0 +1,2 @@
+// Separate file to generate the correct output file name in the dist folder
+export * from './utils';
