+++ frontend-react/src/lib/mod-widget.ts
@@ -1,0 +1,2 @@
+// Separate file to generate the correct output file name in the dist folder
+export * from './widget';
