+++ frontend-react/src/lib/user-data/index.ts
@@ -1,2 +1,12 @@
+/**
+ * @packageDocumentation
+ *
+ * Types and utilities for interacting with the user data stored on the user's device.
+ *
+ * @example
+ * ```tsx
+ * import { ... } from '@wuespace/telestion/user-data';
+ * ```
+ */
 export * from './model.ts';
 export * from './state.ts';
