+++ frontend-react/src/lib/utils/index.ts
@@ -1,0 +1,13 @@
+/**
+ * @packageDocumentation
+ *
+ * Various utility functions around building Telestion frontend apps.
+ *
+ * @example
+ * ```ts
+ * import { ... } from '@wuespace/telestion/utils';
+ * ```
+ */
+export * from './promise-utils.ts';
+export * from './user-data-utils.ts';
+export * from './file-utils.ts';
