+++ frontend-react/src/lib/widget/index.ts
@@ -1,3 +1,13 @@
+/**
+ * @packageDocumentation
+ *
+ * Everything you need for building and/or displaying widgets.
+ *
+ * @example
+ * ```tsx
+ * import { ... } from '@wuespace/telestion/widget';
+ * ```
+ */
 export * from './model.ts';
 export * from './state.ts';
 export * from './component/widget-renderer.tsx';
